feat: use explicit 64-bit unsigned int for decoded size

+++ crates/compression-codecs/src/lib.rs
@@ -100,5 +100,5 @@ pub trait Decode {
 
 pub trait DecodedSize {
     /// Returns the size of the input when uncompressed.
-    fn decoded_size(input: &[u8]) -> Result<usize>;
+    fn decoded_size(input: &[u8]) -> Result<u64>;
 }
